% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newproject.R
\name{newproject}
\alias{newproject}
\title{Creates the standard Problem Forward infrastructure for a new project.}
\usage{
newproject()
}
\description{
After creating a repo on GitHub and cloning it to your session, use this
function to create the basic repo infrastructure for a project in that
project's working directory.
}
\examples{
TBD.
}
