% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_pf.R
\name{theme_pf}
\alias{theme_pf}
\title{Applies the Problem Forward ggplot2 theme.}
\usage{
theme_pf(base_size = 10, font = NA, discrete = TRUE, grid = "XY")
}
\arguments{
\item{base_size}{base font size for the plot; defaults to \code{10}}

\item{font}{font family to use; defaults to \code{"Roboto"}}

\item{discrete}{use a discrete color palette; defaults to \code{TRUE}}

\item{grid}{major axis grid lines in plot; defaults to \code{"XY"}.
options include: \code{"X"} for only X grid lines,
                 \code{"Y"} for only Y grid lines,
                 \code{"XY"} for both.}
}
\description{
You can \code{+ theme_pf()} to an existing ggplot to apply our branding.
It should be one of the first things you do after adding in the ggplot
\code{aes} and \code{geom}, so that you can fiddle with the rest of the
layout afterward.
}
\examples{
TBD.
}
